Refactor lib initialisation
This is perhaps an intermediary solution.

- Most itinialisation elements are now defined in a Config type;
- Additional brush and console can be defined through a modifier
  function given as parameter to boot;

The latter is not super user friendly and will look weird if a lot of
brushes / consoles must be displayed however.

Additionally, this forced me to handle better brush access on rendering,
but it is still a bit unnatural in the way it is designed.