Improve support for changing tilesizes
It became obvious that we need to be able to properly switch tilesizes
(text UI typically wants characters that are less square). So I embarked
on a big refactoring.

- We now ensure that the main draw function is done to target different
 consoles (with different layout computations bases for each ones).
- This lead to a better component drawing context, with a more stateful
 design which simplifies a bit their definition (no need to pass a
display context around, and its stateful nature is now clear).

We're still in awkwards places at time, particularly when we need to
switch brush, but I've decided for a design that I find a bit more
declarative. There is now more state, that's for sure, but the design is
hopefully clearer.

Finally, it adds utility to fill a console with a background colour, and
fixes a bug where user-defined consoles were never positioned properly
due to a wrong initialisation of the drawing context.