Add extent reporting à la Brick
The previous commit gave us an API for viewport, but forced users to
compute manually the size of a component. We definitely do not want
that.

Since we have potentially dynamically sized components, there is no
other way around but to collect their actual size, like Brick does. I've
borrowed also their terminology as it is very clear and useful.

The only constraint is that now interactive components should report
their name so it can get stored. This is only done for viewports right
now, but I think the next commit will generalize it to lists, buttons
and text inputs so they benefit from extent collection. As a _very_ good
bonus, this will make implementing mouse support very easy for all of
these.