Add rotation and flipping
It took a while to understand why the initial demo was weird. Of course,
non square tilesets won't rotate well... it could be solved by swapping
the rect dimension on rotation, but honestly, it's not worth the effort;
rotations are mostly interesting for tilesets (and for grid display in
general; non-square tilesets are more a ui thing).