Add support for overlay and animation
These are two eye-candy things that were missing, are not too hard to
implement, and can really improve on game quality.

A cool demo with a torch-like effect was added. It's not perfect (light
goes "inside of the wall", so it's not super realistic), but good enough
for engine demonstration purpose. I'm also happy because it shows you
can use something else than `Array` to store the map.